Refactor the brew command into one file per command
We're scared to refactor Homebrew because the project runs on top of Git and
we encourage you to hack on the innards. I dread to think how many merge
conflicts this commit will generate. But ultimately the original design was
stupid and we need to split out the code across many more files to help us
poor humans categorise the code better so that we don't get lost and make
mistakes.

The refactor is minimal.

If you merge you will need to pay attention to brew.h.rb (as it is deleted)
and bin/brew (as command logic is gone). It will be painful, but you will just
have to help git out by moving any changes around manually.

Note compatibility.rb. It ensures that any function renames or removals don't
break anything. We're pretty serious about backwards compatibility. And that's
because we encourage you to hack around with the innards. And we couldn't do
that if we would then just make stuff disappear behind your back.